% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldup.r
\name{alldup}
\alias{alldup}
\title{Show all duplicate entries in a data frame/table}
\usage{
alldup(DF, by)
}
\arguments{
\item{DF}{A data frame}

\item{by}{The column of \code{DF} you want to look at duplicates of}
}
\description{
Not just the "duplicated" ones
}
\examples{
DF <- data.frame(x = 1:3, y = c(1, 1, 2))
alldup(DF, "x")
alldup(DF, "y")
}
