% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaCol.r
\name{alphaCol}
\alias{alphaCol}
\title{Take a solid color and turn it semi-transparent}
\usage{
alphaCol(col, alpha)
}
\arguments{
\item{col}{Color}

\item{alpha}{Semi-transparency value}
}
\description{
Superseded by \code{scales::alpha()}
}
\examples{
alphaCol("red", 0.5)

}
