% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attach.R
\name{Attach}
\alias{Attach}
\title{Detach before attaching}
\usage{
Attach(x, name = deparse(substitute(x)), ...)
}
\description{
Detach before attaching
}
\examples{
Attach(airquality)
Attach(airquality)   # Ordinarily this would provoke warnings

}
