% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.bayes.R
\name{binom.bayes}
\alias{binom.bayes}
\title{Bayesian analysis of one-sample binomial data using conjugate beta priors}
\usage{
binom.bayes(
  x,
  n,
  a = 1,
  b = 1,
  level = 0.95,
  null,
  plot = FALSE,
  add = FALSE,
  xlab = "p",
  ylab = "Posterior density",
  col = "blue",
  ...
)
}
\arguments{
\item{x}{Number of "successes"}

\item{n}{Number of trials}

\item{a}{'alpha' parameter for beta prior. Default: uniform prior}

\item{b}{'beta' parameter for beta prior. Default: uniform prior}

\item{level}{For posterior interval; .95 for a 95\% credible interval}

\item{null}{Point null to evaluate as a kind of hypothesis test}

\item{plot}{Draw a plot?  Default: false}

\item{add}{Add to existing plot?  Default: false}

\item{xlab}{xlab for plot}

\item{ylab}{ylab for plot}

\item{col}{color of density line for plot}

\item{...}{Additional arguments to \code{plot()}}
}
\description{
Bayesian analysis of one-sample binomial data using conjugate beta priors
}
\examples{
binom.bayes(10, 16)
binom.bayes(0, 21, plot=TRUE)

}
