% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.r
\name{Boot}
\alias{Boot}
\title{Simplified interface to boot package for basic statistics}
\usage{
Boot(data, statistic, R, ...)
}
\arguments{
\item{data}{Data to bootstrap; can be a vector or something matrix-like}

\item{statistic}{Statistic to be applied to data}

\item{R}{Number of bootstrap replicates}

\item{...}{Additional arguments to \code{boot()}}
}
\description{
Simplified interface to boot package for basic statistics
}
\examples{
x <- rexp(100)
Boot(x, mean, 999)

}
