% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breheny.r
\docType{package}
\name{breheny-package}
\alias{breheny}
\alias{breheny-package}
\title{breheny: Miscellaneous functions that Patrick Breheny finds useful}
\description{
Miscellaneous functions that Patrick Breheny finds useful
}
\author{
\strong{Maintainer}: Patrick Breheny \email{patrick-breheny@uiowa.edu}

}
