% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchWarning.R
\name{catchWarning}
\alias{catchWarning}
\title{Catch warnings}
\usage{
catchWarning(expr)
}
\description{
Catch warnings
}
\examples{
out <- catchWarning(log('a'))
out$value
out <- catchWarning(as.numeric(c('1', 'a')))
out
out <- catchWarning(1+1)
out
}
