% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{Quick inspection of columns in a data frame / table for weirdness}
\usage{
checkData(Data)
}
\arguments{
\item{Data}{A data frame (or data table)}
}
\description{
Provides histograms and ranges for numeric features, and tables for categorical features, along with missing data information.
}
\examples{
checkData(airquality)

}
