% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_man.R
\name{chk_man}
\alias{chk_man}
\title{Check code/documentation agreement using \code{devtools::check_man()}}
\usage{
chk_man()
}
\description{
Intended to be used as an Rstudio addin (which is why it doesn't accept
arguments).
}
