% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_int.R
\name{ci_int}
\alias{ci_int}
\title{Construct confidence intervals in the presence of interactions}
\usage{
ci_int(fit, term, ...)
}
\arguments{
\item{fit}{A model fit, as produced by \code{\link[=lm]{lm()}}, \code{\link[=glm]{glm()}}, \code{\link[survival:coxph]{survival::coxph()}}, etc.}

\item{term}{Term of interest}

\item{...}{Term to condition on; ideally a vector. See examples.}
}
\value{
A \code{data.frame} suitable for use in \code{\link[=ci_plot]{ci_plot()}}
}
\description{
Construct confidence intervals in the presence of interactions
}
\examples{
fit <- lm(Ozone ~ Solar.R + Wind * Temp, airquality)
ci_int(fit, Wind, Temp=c(70, 80, 90))
ci_int(fit, Wind, Temp=c(70, 80, 90)) |> ci_plot()
airquality$Heat <- cut(airquality$Temp, 3, labels=c("Cool","Mild","Hot"))
fit <- lm(Ozone ~ Solar.R + Wind * Temp, airquality)
ci_int(fit, Wind, Temp) |> ci_plot()
}
\seealso{
\code{\link[=ci_plot]{ci_plot()}}
}
