% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_ipd.r
\name{ci_ipd}
\alias{ci_ipd}
\title{Given a confidence interval, simulate individual patient data}
\usage{
ci_ipd(ci, n, df = Inf)
}
\arguments{
\item{ci}{Confidence interval}

\item{n}{Sample size}

\item{df}{Degrees of freedom (default: infinite)}
}
\value{
A numeric vector of individual patient data
}
\description{
Given a confidence interval, simulate individual patient data
}
\examples{
ci_ipd(c(-0.31, -0.02), 10)
}
