% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_plot.R
\name{ci_plot}
\alias{ci_plot}
\alias{ci_plot.matrix}
\title{Create forest plot of confidence intervals}
\usage{
ci_plot(obj, ...)

\method{ci_plot}{matrix}(
  obj,
  sort = TRUE,
  diff = (ncol(obj) == 4),
  null = 0,
  trans,
  p_label = FALSE
)
}
\arguments{
\item{obj}{The object to be plotted; can be a matrix of raw values or a model object}

\item{...}{Not used}

\item{sort}{Sort parameters by estimate? (default: true)}

\item{diff}{Include tests of difference / p-values?}

\item{null}{Draw a line representing no effect at this value (default: 0)}

\item{p_label}{Label p-values (p=0.02 instead of just 0.02)? (default: FALSE)}
}
\description{
"Forest plot"-style plotting of confidence intervals from a regression model. Basic input is a matrix with columns of estimate/lower/upper, along with an optional 4th column for the p-value. Also works with a variety of models (lm/glm/coxph/etc).
}
\details{
NEEDS TAU AND PROPER RHL FOR PLABEL
}
\examples{
# Supplying a matrix
B <- cbind(1:3, 0:2, 2:4)
rownames(B) <- LETTERS[1:3]
ci_plot(B)

# Supplying a fitted model object
fit <- lm(Ozone ~ Solar.R + Wind + Temp, airquality)
ci_plot(fit)
}
