% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciband.r
\name{ciband}
\alias{ciband}
\title{Add a confidence interval band to a KM curve}
\usage{
ciband(obj, ...)
}
\arguments{
\item{obj}{A \code{survfit} or \code{survfitms} object}

\item{...}{Not used}
}
\description{
Add a confidence interval band to a KM curve
}
\examples{
library(survival)

fit <- survfit(Surv(time, status) ~ 1, veteran)
Plot(fit)
ciband(fit)

fit <- survfit(Surv(time, status) ~ celltype, veteran)
Plot(fit)
ciband(fit)

}
