% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_share.R
\name{cloud_share}
\alias{cloud_share}
\title{Share Amazon CloudFront file}
\usage{
cloud_share(file, out = stdout(), expiration = 365, r)
}
\arguments{
\item{file}{Character vector of files to share.If path is included, it will be included in the download script save location.}

\item{out}{A connection object or character string, as in \verb{[writeLines()]}, to which R commands are to be saved.  If nothing is supplied, output is to console.}

\item{expiration}{How long until URLs expire (in days)? Numeric.  Default: 365.}

\item{r}{Output R code (as opposed to link)? By default, the function will return links if \code{out=stdout()} and R code if \code{out} otherwise. However, this can be overridden by explicitly setting r to be \code{TRUE} or \code{FALSE}.}
}
\value{
Either a url or R script. The latter is particularly useful for batch downloads and specifying file save paths.
}
\description{
Shares Amazon CloudFront file. Requires aws command line functionality and quite a bit of setup; see bin/cloud-share.
}
\examples{
\dontrun{
s3_share('~/web/pdf/BrehenyCV.pdf')
s3_share(c('~/web/pdf/BrehenyCV.pdf', '~/web/pdf/mnet.pdf'), r=TRUE)
}
}
