% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table.R
\name{contingency_table}
\alias{contingency_table}
\title{Format a contingency table for use with knitr}
\usage{
contingency_table(Tab, digits = 2)
}
\arguments{
\item{Tab}{Number of random strings to generate}

\item{digits}{Number of characters in each random string}
}
\description{
Format a contingency table for use with knitr
}
\examples{
Tab <- table(A=rnorm(100)>0, B=rnorm(100)>0)
contingency_table(Tab)

}
