% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_tree.r
\name{cv_tree}
\alias{cv_tree}
\alias{plot.cv_tree}
\title{Cross-validation for regression trees}
\usage{
cv_tree(formula, data, ..., nfolds = 10, seed, p = seq(0.01, 0.2, 0.01))

\method{plot}{cv_tree}(x, rev = FALSE, ...)
}
\arguments{
\item{formula}{Model formula, as in \code{ctree()}}

\item{data}{Data frame containing the variables in the model, as in \code{ctree()}}

\item{...}{Additional arguments to \code{ctree()}}

\item{nfolds}{Number of cross-validation folds.  Default: 10.}

\item{seed}{Seed for reproducibility}

\item{p}{Complexity parameter; see \code{mincriterion} in \code{ctree_control()}}

\item{x}{Result of a call to}

\item{rev}{Reverse horizontal axis (i.e., plot from high-to-low complexity)?  Defualt: FALSE.}
}
\description{
Cross-validation for regression trees
}
\examples{
cv <- cv_tree(mpg ~ ., mtcars)
plot(cv)

}
