% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.r
\name{density_plot}
\alias{density_plot}
\title{Density plot}
\usage{
density_plot(M, labs)
}
\arguments{
\item{M}{A vector or matrix of values upon which the density plot is based; if a matrix, separate densities will be estimated and plotted for each column}

\item{labs}{If supplied, a legend is printed; order of \code{labs} should correspond to columns of \code{M}}
}
\description{
Like \code{dnplot()}, but using ggplot2.
}
\examples{
M <- cbind(rgamma(1000, 2, 2), rnorm(1000))
colnames(M) <- c("Gamma", "Normal")
density_plot(M[, 1])
density_plot(M)
density_plot(rexp(100, 2), labs = "Exponential")
density_plot(data.frame(Gamma = rgamma(1000, 2, 2), Normal = rnorm(1000)))
}
