% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devload.r
\name{devload}
\alias{devload}
\title{Load a package from its dev/ location (from source, not installed)}
\usage{
devload(pkg)
}
\arguments{
\item{pkg}{Package name (character)}
}
\description{
Load a package from its dev/ location (from source, not installed)
}
\examples{
\dontrun{
devload(ncvreg)
}
}
