% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{dict}
\alias{dict}
\title{Create a key-value dictionary}
\usage{
dict(...)
}
\arguments{
\item{...}{Either supply a single, unnamed, two-column table or a list of
\code{key=value} pairs; see examples. If a two-column table, first
column has the keys.}
}
\description{
Keys must be valid
}
\examples{
d <- dict(a = 5, b = 10)
d$b
d <- data.frame(letters[1:10], 1:10) |>
  dict()
d$f
}
