% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnplot.r
\name{dnplot}
\alias{dnplot}
\title{Density plot}
\usage{
dnplot(
  M,
  labs,
  pos = FALSE,
  lwd = 3,
  bty = "n",
  lty = 1,
  col = pal(p),
  ylab = "Density",
  xlab = "",
  las = 1,
  dens.par = NULL,
  ...
)
}
\arguments{
\item{M}{A vector or matrix of values upon which the density plot is based; if a matrix, separate densities will be estimated and plotted for each column}

\item{labs}{If supplied, a legend is printed; order of \code{labs} should correspond to columns of \code{M}}

\item{pos}{Are the densities strictly positive?  If so, a boundary correction is applied.  Default: FALSE}

\item{lwd, bty, lty, col, xlab, ylab, las}{Plotting options, as in \code{par()}}

\item{dens.par}{Additional arguments to \code{density()}}

\item{...}{Additional arguments to \code{plot()}}
}
\description{
Primarily intended for plotting posterior densities
}
\examples{
M <- cbind(rgamma(1000, 2, 2), rnorm(1000))
dnplot(M[,1])
dnplot(M[,1], pos=TRUE)
dnplot(M)
dnplot(M, labs=c('Gamma', 'Normal'))
dnplot(M, labs=c('Gamma', 'Normal'))
}
