% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{download_data}
\alias{download_data}
\alias{read_data}
\title{Download data and read into R}
\usage{
download_data(name)

read_data(name)
}
\arguments{
\item{name}{Name of data set to read/download}
}
\description{
These functions are downloading and reading data sets from my courses into an
R session. To keep the size of the package manageable, these data sets are
not included with the package and must be downloaded separately using
\code{download_data()}; this only needs to be done once.
}
\examples{
download_data("anorexia") # Download a specific data set
dat <- read_data(whickham)
head(dat)
}
