% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.r
\name{downloadData}
\alias{downloadData}
\alias{readData}
\title{Download data and read into R}
\usage{
downloadData(name)

readData(name)
}
\arguments{
\item{name}{Name of data set to read}
}
\description{
These functions are downloading and reading data sets from my courses into an R session.
To keep the size of the package manageable, these data sets are not included with the package and must be downloaded separately using \code{downloadData()}; this only needs to be done once.
}
\examples{
\dontrun{
downloadData()         # Download all data sets
downloadData(bcTCGA)   # Download a specific data set}
\dontrun{
Data <- readData(whickham)
head(Data)}
}
