% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropMiss.r
\name{DropMiss}
\alias{DropMiss}
\title{Drop missing columns from a data table}
\usage{
DropMiss(DT)
}
\arguments{
\item{DT}{The data table}
}
\value{
A data table; no columns will be entirely missing
}
\description{
Currently only works for data tables, and only for completely missing columns
}
\examples{
DT <- data.table::data.table(A = 1:3, B = NA, C = c(runif(2), NA))
DropMiss(DT)

}
