% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epower.R
\name{epower}
\alias{epower}
\alias{esamsize}
\title{Power and sample size for time-to-event data based on exponential closed form solution}
\usage{
epower(n, r, alpha = 0.05, w = c(1, 1), n1, n2, lam, cens)

esamsize(r, w = c(1, 1), alpha = 0.05, power = 0.8, lam, cens)
}
\arguments{
\item{n}{The table}

\item{r}{Effect size (ratio of hazards/medians)}

\item{power}{Desired power; default: 0.8}
}
\description{
Power and sample size for time-to-event data based on exponential closed form solution
}
\section{Functions}{
\itemize{
\item \code{epower}: Calculate power

\item \code{esamsize}: Calculate sample size
}}

\examples{
epower(30, 1.5)
esamsize(1.5)
}
