% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_chart.r
\name{event_chart}
\alias{event_chart}
\title{Create a time-to-event chart using ggplot}
\usage{
event_chart(Data, shift = FALSE, scale = 1)
}
\arguments{
\item{Data}{A data frame or data.table of dates, organized in the following way (the first two columns are plotted with a line segment, the remaining columns are plotted with a symbol):
\itemize{
\item Column 1: Follow-up start date
\item Column 2: Follow-up end date
\item Additional columns: Additional events that may occur during the follow-up period
}}

\item{shift}{Shift time so that 0 represents start of follow-up for everyone? (default: FALSE)}

\item{scale}{If shifted, what units? (default: scale=1, meaning days)}
}
\description{
Inspired by \code{Hmisc::event.chart()}
}
\examples{
df <- read.csv(system.file("extdata/cdcaids.txt", package = "breheny"))
event_chart(df, FALSE)
event_chart(df, TRUE)
event_chart(df, TRUE, 365.25)
}
