% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2num.R
\name{factor2num}
\alias{factor2num}
\title{Convert factor to number}
\usage{
factor2num(f)
}
\arguments{
\item{f}{Factor with levels like 20, 40, 60, that are simply numbers}
}
\description{
Convert factor to number
}
