% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flowchart}
\alias{flowchart}
\alias{flow}
\alias{addNode}
\alias{plot.flow}
\title{Create a flowchart}
\usage{
addNode(obj, node, parent, edge = "")

\method{plot}{flow}(obj, x, xm = 0.1, ym = 0.1)
}
\arguments{
\item{obj}{\code{flow} object; in addNode, a new object will be created if not supplied}

\item{node}{Text for new node}

\item{parent}{Parent of new node}

\item{edge}{???}

\item{x}{left/right positioning of flowchart elements}

\item{xm}{x (horizontal) margin (.2 = 20\% whitespace on left/right margins)}

\item{ym}{y (vertical) margin (.2 = 20\% whitespace on top/bottom margins)}
}
\description{
Create a flowchart
}
\note{
Probably would be better with R6/RC object orientation
}
\examples{
X <- addNode("All children\\n(n=935)")
addNode(X, "Returned form\\n(n=511)", parent=1)
addNode(X, "Baseline: 0\\n(n=447)", parent=2)
addNode(X, "Baseline: 1 or 2\\n(n=14)", parent=2)
addNode(X, "Baseline: 3\\n(n=50)", parent=2)
addNode(X, "Initiation\\n(n=315)", parent=3)
addNode(X, "Completion\\n(n=276)", parent=4)
addNode(X, "Completion\\n(n=12)", parent=6)
plot(X, c(0, 0, -1, 0, 1, -1, 0, -1), xm=.2)
}
