% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbytes.r
\name{format_object_size}
\alias{format_object_size}
\title{This is literally copied from utils:::format.object_size}
\usage{
format_object_size(x, units = "b", standard = "auto", digits = 1L, ...)
}
\arguments{
\item{x, units, standard, digits, ...}{See \code{utils:::format.object_size()}}
}
\description{
This is literally copied from utils:::format.object_size
}
