% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.r
\name{format_p}
\alias{format_p}
\title{Format p-values nicely}
\usage{
format_p(p, digits = 2, label = FALSE)
}
\arguments{
\item{p}{Vector of p-values to format}

\item{digits}{Digits after decimal point to display}

\item{label}{If \code{label=TRUE}, prints \code{p=0.01}; otherwise, just returns \code{0.01}}
}
\description{
Format p-values nicely
}
\examples{
format_p(0.000003434)
format_p(0.12)
format_p(0.12, label = TRUE)
format_p(c(0.12, 0.000003434)) |> cat(sep = "\n")
format_p(c(0.12, 0.000003434), label = TRUE) |> cat(sep = "\n")
}
