% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatP.r
\name{formatP}
\alias{formatP}
\title{Format p-values nicely}
\usage{
formatP(p, digits = 2, label = FALSE)
}
\arguments{
\item{p}{Vector of p-values to format}

\item{digits}{Digits after decimal point to display}

\item{label}{If \code{label=TRUE}, prints \code{p=0.01}; otherwise, just returns \code{0.01}}
}
\description{
Format p-values nicely
}
\examples{
formatP(0.000003434)
formatP(0.12)
formatP(0.12, label=TRUE)

}
