% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Generates a random design matrix and outcome.}
\usage{
genData(n, J, K = 1, beta, family = c("gaussian", "binomial"),
  J0 = ceiling(J/2), K0 = K, SNR = 1, sig = c("homogeneous",
  "heterogeneous"), sig.g = c("homogeneous", "heterogeneous"), rho = 0,
  rho.g = rho, corr = c("exchangeable", "autoregressive"))
}
\arguments{
\item{n}{Sample size}

\item{J}{Number of groups}

\item{K}{Number of explanatory variables per group}

\item{beta}{Vector of regression coefficients in the generating model, or, if a scalar, the value of each nonzero regression coefficient.}

\item{family}{Generate \code{y} according to linear (\code{"gaussian"})
or logistic (\code{"binomial"}) model}

\item{J0}{Number of nonzero groups}

\item{K0}{Number of nonzero coefficients per group}

\item{SNR}{Signal to noise ratio}

\item{sig}{Should the groups be heterogeneous (in beta) or homogeneous}

\item{sig.g}{Should the coefficients within a group be heterogeneous or homogeneous}

\item{rho}{Correlation between groups}

\item{rho.g}{Correlation between parameters within a group}
}
\description{
This function generated a random design matrix \code{X} and
response vector \code{y} for use in regression testing/simulation.
}
