% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-description.r
\name{generic_description}
\alias{generic_description}
\title{Copies a generic DESCRIPTION file to current working directory}
\usage{
generic_description()
}
\description{
Copies a generic DESCRIPTION file to current working directory
}
