% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbar.r
\name{ggbar}
\alias{ggbar}
\title{Quick wrapper for a gg bar plot}
\usage{
ggbar(x)
}
\arguments{
\item{x}{A named vector}
}
\value{
A gg object
}
\description{
Quick wrapper for a gg bar plot
}
\examples{
x <- rexp(10)
names(x) <- letters[1:10] |> sample()
ggbar(x)
}
