% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbox.r
\name{ggbox}
\alias{ggbox}
\title{Boxplot of a data frame / matrix using ggplot}
\usage{
ggbox(X, xlab)
}
\arguments{
\item{X}{Matrix of values}

\item{xlab}{If omitted, no space will be allocated for x axis title (ylab
can be specified as usual; see examples)}
}
\description{
Works with labels, as from the \code{labelled} package. Axis labels are handled
somewhat differently depending on whether a single box is present or multiple.
}
\examples{
ggbox(matrix(rnorm(50), 10, 5))
ggbox(matrix(rnorm(50), 10, 5), xlab='categories')
ggbox(rexp(10))
ggbox(rexp(10), 'A plot of') + ggplot2::ylab('Something')

# Using labels
X <- iris[, 1:4]
attr(X[[4]], 'label') <- "The petal's width"
ggbox(X)
ggbox(X[,4])
}
