% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcontour.r
\name{ggcontour}
\alias{ggcontour}
\title{Wrapper to create geom_contour of a function}
\usage{
ggcontour(x, y, f)
}
\arguments{
\item{x, y}{Vectors defining the x and y axis}

\item{f}{The function to plot}
}
\value{
A \code{gg} object
}
\description{
Wrapper to create geom_contour of a function
}
\examples{
x <- y <- seq(-10, 10, 0.1)
f <- function(x, y) { r <- sqrt(x^2+y^2); 10 * sin(r)/r }
ggcontour(x, y, f)
}
