% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{Simple interface to ggplot2 histogram}
\usage{
gghist(x, n = 10)
}
\arguments{
\item{x}{A vector of values to make a histogram out of}

\item{n}{Number of bins}
}
\description{
Simple interface to ggplot2 histogram
}
\examples{
Hist(rnorm(100))
}
