% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmat.r
\name{ggmat}
\alias{ggmat}
\title{Line plot of a matrix using ggplot}
\usage{
ggmat(x, y)
}
\arguments{
\item{x, y}{A vector (\code{x}) and matrix (\code{y}); if only x is provided, it is
taken to be the matrix of responses and \code{1:n} used as the horizontal axis}
}
\description{
Essentially a replacement for matplot.
}
\examples{
sines <- outer(1:20, 1:4, function(x, y) sin(x / 20 * pi * y))
ggmat(sines)
ggmat((1:20)^2, sines)
}
