% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbytes.R
\name{hbytes}
\alias{hbytes}
\title{Print bytes in nice human-readable format}
\usage{
hbytes(x)
}
\arguments{
\item{x}{Number of bytes (integer)}
}
\description{
Basically format.object_size from utils, but vectorized and includes .0
}
\examples{
hbytes(12300)
hbytes(c(12300, 12345))
}
