% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist.r
\name{Hist}
\alias{Hist}
\title{Nicer-looking histogram}
\usage{
Hist(
  x,
  ...,
  col = "gray",
  border = "white",
  n = 10,
  breaks = seq(min(x, na.rm = TRUE), max(x, na.rm = TRUE), len = n),
  main = "",
  las = 1,
  xlab = deparse(substitute(x))
)
}
\arguments{
\item{x}{A vector of values to make a histogram out of}

\item{...}{Further arguments to \code{hist()}}

\item{col}{Fill color for histogram bars; default: gray}

\item{border}{Border (outline) of histogram bars; default: white}

\item{n}{Number of bins}

\item{breaks}{Specific locations at which bins are to be constructed; overrides \code{n}}

\item{main}{Main title; default: none}

\item{las}{See \code{par()}}

\item{xlab}{Just like in every other plot}
}
\description{
Nicer-looking histogram
}
\examples{
Hist(rnorm(100))

}
