% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_normal.r
\name{inverse_normal}
\alias{inverse_normal}
\title{Apply inverse normal rank transformation}
\usage{
inverse_normal(x)
}
\arguments{
\item{x}{Vector of numeric values}
}
\description{
Apply inverse normal rank transformation
}
\examples{
c(-5:5, NA, NA, NA) |> inverse_normal()
}
