% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotl.R
\encoding{UTF-8}
\name{likelihood_plot}
\alias{likelihood_plot}
\alias{plotL}
\alias{plotl}
\title{Plot likelihood function}
\usage{
plotL(x, l, xlab = expression(theta), ylab = expression(L(theta)),
  bty = "n", ...)

plotl(x, l, xlab = expression(theta), ylab = expression("\u2113" *
  (theta)), bty = "n", ...)
}
\arguments{
\item{x}{Parameter values (on x axis)}

\item{l}{Likelihood / log-likelihood (on y axis)}

\item{xlab}{Default: theta}

\item{ylab}{Default: L(theta)}

\item{bty}{Default: no boundary box}

\item{...}{Further arguments to \code{plot()}}
}
\description{
\code{plotL()} plots the likelihood, \code{plotl()} plots the log-likelihood
}
\examples{
x <- seq(-3, 4, length=99)
L <- dnorm(1, x)
l <- dnorm(1, x, log=TRUE)
plotL(x, L)
plotl(x, l)
}
