% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique.r
\name{make_unique}
\alias{make_unique}
\title{Ensure all entries in a string are uniquely labeled}
\usage{
make_unique(x, sep = "_")
}
\arguments{
\item{x}{A character vector or something coerceable to a character vector.}

\item{sep}{Separator between label and suffix.  Default: _.}
}
\description{
Leaves entries untouched if they are already unique; adds a suffix like '_1', etc. if necessary.
}
\examples{
make_unique(LETTERS[c(1, 1, 2, 2, 2, 3, 4)])

}
