% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_dt.r
\name{matrix_to_dt}
\alias{matrix_to_dt}
\title{Convert matrix to data.table}
\usage{
matrix_to_dt(x, row = "row", col = "col", val = "val")
}
\arguments{
\item{x}{The matrix}

\item{row}{Name for row variable (default: 'row')}

\item{col}{Name for col variable (default: 'col')}

\item{val}{Name for val variable (default: 'val')}
}
\description{
Convert matrix to data.table
}
\examples{
x <- matrix(runif(9), 3, 3)
rownames(x) <- letters[1:3]
colnames(x) <- letters[4:6]
matrix_to_dt(x)
matrix_to_dt(x, "sample", "gene", "expression")
}
