% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.survfit.R
\name{median.survfit}
\alias{median.survfit}
\title{Calculate median survival time of a survfit object}
\usage{
\method{median}{survfit}(x, ...)
}
\arguments{
\item{x}{A survfit object}

\item{...}{Not used}
}
\description{
Based on survmean function from survival package
}
\examples{
library(survival)
fit <- survfit(Surv(time, status) ~ x, data = aml)
median(fit)

}
