% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcolor.r
\name{mixcolor}
\alias{mixcolor}
\title{Mix two colors}
\usage{
mixcolor(x, y, frac = 0.5, alpha = 1)
}
\arguments{
\item{x, y}{Colors (anything accepted by \code{col2rgb()})}

\item{frac}{Mixing fraction, proportion of \code{x} (default: 0.5)}

\item{alpha}{Transparency (default: 1)}
}
\description{
Mix two colors
}
\examples{
mixcolor("#008DFFFF", "white")
mixcolor("#008DFFFF", "white", frac = 0.9)
mixcolor("#008DFFFF", "white", alpha = 0.2)
}
