% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normProb.r
\name{normProb}
\alias{normProb}
\title{Draw a normal curve and shade area under an interval (probability)}
\usage{
normProb(a, b, outside = FALSE, ...)
}
\arguments{
\item{a, b}{Endpoints of the interval}

\item{outside}{Shade the area outside (a,b)?  Default is FALSE (i.e., shade the area inside a and b)}

\item{...}{Additional arguments to \code{plot()}}
}
\description{
Draw a normal curve and shade area under an interval (probability)
}
\examples{
normProb(1, 2)
normProb(1, 2, outside = TRUE)
normProb(1, 2, lwd = 5, col = "pink", main = "Nice!")
}
