% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrisk.r
\name{nrisk}
\alias{nrisk}
\title{Add an axis describing number at risk to a KM curve}
\usage{
nrisk(fit, cex = 0.8)
}
\arguments{
\item{fit}{A \code{survfit} or \code{survfitms} object}

\item{cex}{Character expansion parameter for axis labels, as in \code{par()}}
}
\description{
Add an axis describing number at risk to a KM curve
}
\examples{
library(survival)

fit <- survfit(Surv(time, status) ~ 1, veteran)
Plot(fit)
nrisk(fit)

fit <- survfit(Surv(time, status) ~ celltype, veteran)
op <- par(mar=c(5,5,6,7))
Plot(fit, legend='right')
nrisk(fit)
par(op)

}
