% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_url.r
\name{open_url}
\alias{open_url}
\title{Open a url: quietly, in the background, and in a new window}
\usage{
open_url(x)
}
\arguments{
\item{x}{URL to open}
}
\description{
Open a url: quietly, in the background, and in a new window
}
\examples{
\donttest{
open_url('www.google.com')
}
}
