% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_aux.r
\name{parse_aux}
\alias{parse_aux}
\title{Parse latex .aux files for fig/tab/etc. numbers}
\usage{
parse_aux(aux, id = c("Fig", "Tab", "Ex", "ex"))
}
\arguments{
\item{aux}{.aux files (or a vector of .aux files)}

\item{id}{Function assumes labels are of form category:label; \code{id} is a character vector of categories to search for}
}
\description{
Parse latex .aux files for fig/tab/etc. numbers
}
\examples{
\dontrun{
parse_aux("file.aux")
parse_aux(c("ch1.aux", "ch2.aux"), "ex")
}

}
