% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.r
\name{pca_plot}
\alias{pca_plot}
\title{Make a PCA/tSNE plot}
\usage{
pca_plot(
  X,
  grp,
  txt = FALSE,
  tsne = FALSE,
  perp = 30,
  dims = 2,
  gg = TRUE,
  ellipse = FALSE,
  legend,
  plot = TRUE,
  ...
)
}
\arguments{
\item{X}{A numeric matrix}

\item{grp}{Optional grouping factor to color instances by}

\item{txt}{Use rownames as labels instead of dots (default: false)}

\item{tsne}{Use tSNE instead of PCA?  (default: false)}

\item{perp}{Perplexity (for tSNE)}

\item{dims}{Dimensions to return (for tsne)}

\item{gg}{Use ggplot2? (default: true)}

\item{ellipse}{Draw mvn-ellipses around groups? (default: false, only available for ggplot)}

\item{legend}{Include a legend (default: true)}

\item{plot}{Create plot? (default: true)}

\item{...}{Further arguments to \code{plot()}}
}
\value{
If base plot, returns PCA object invisibly. If ggplot, returns the plot object.
}
\description{
Make a PCA/tSNE plot
}
\examples{
pca_plot(iris[,1:4])
pca_plot(iris[,1:4], grp=iris[,5])
pca_plot(iris[,1:4], grp=iris[,5], ellipse=TRUE)
pca_plot(iris[,1:4], grp=iris[,5], tsne=TRUE)
pca_plot(USArrests, txt=TRUE)
pca_plot(USArrests, txt=TRUE, grp=rep(LETTERS[1:5], each=10))
}
