% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot}
\alias{Plot}
\alias{Plot.BinaryTree}
\alias{Plot.rpart}
\alias{Plot.survfit}
\title{Customized plotting functions}
\usage{
Plot(obj, ...)

\method{Plot}{BinaryTree}(obj, pval = FALSE, summary = FALSE, digits = 1, ...)

\method{Plot}{rpart}(obj, ...)

\method{Plot}{survfit}(
  obj,
  legend = c("top", "right", "none"),
  xlab = "Time",
  ylab = "Survival",
  conf.int = FALSE,
  col,
  ...
)
}
\arguments{
\item{obj}{Object to be plotted}

\item{...}{Additional arguments to underlying plot function}

\item{pval}{Show p-values?  Default=FALSE}

\item{legend}{Where to put the legend. Either 'top', 'right', or 'none'; default: 'top'}
}
\description{
Customized plotting functions
}
\examples{
# KM curves
data(veteran, package='survival')
fit <- survival::survfit(survival::Surv(time, status) ~ celltype, veteran)
Plot(fit)

# Trees
data(kyphosis, package='rpart')
fit <- rpart::rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)
Plot(fit)
fit <- party::ctree(Kyphosis ~ Age + Number + Start, data = kyphosis)

}
