% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.step.R
\name{polygon.step}
\alias{polygon.step}
\title{Draw a filled polygon as a step function}
\usage{
polygon.step(x, y1, y2, border = FALSE, ...)
}
\arguments{
\item{x}{x coordinates; must be one longer than \code{y1} and \code{y2}}

\item{y1}{lower y coordinates}

\item{y2}{upper y coordinates}

\item{border}{Color to draw the border; see \code{polygon()}.  Default: no borders.}

\item{...}{Additional arguments to be passed to \code{polygon()}}
}
\description{
Draw a filled polygon as a step function
}
\examples{
plot(1:10, 1:10, type='n', ylim=c(0,12))
polygon.step(1:10, 0:8, 2:10, col='gray')
lines(1:10, 1:10, type='s')

}
