% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_yaml.r
\name{process_yaml}
\alias{process_yaml}
\title{Split a file into yaml and body}
\usage{
process_yaml(file)
}
\arguments{
\item{file}{Filename (string)}
}
\description{
Split a file into yaml and body
}
\examples{
process_yaml(system.file("extdata/example.rmd", package="breheny"))
}
