% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbox.r
\name{qbox}
\alias{qbox}
\title{Box plot but whiskers show quantiles instead of outliers}
\usage{
qbox(df, x, g, orientation = c("horizontal", "vertical"))
}
\arguments{
\item{df}{A data frame}

\item{x}{The continuous variable}

\item{g}{The categorical (group) variable}

\item{orientation}{Either \code{horizontal} (default) or \code{vertical}}
}
\description{
Box plot but whiskers show quantiles instead of outliers
}
\examples{
qbox(iris, Sepal.Length, Species)
qbox(iris, Sepal.Length, Species, orientation='vertical')
}
