% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcut.r
\name{qcut}
\alias{qcut}
\title{Quantile-based cut}
\usage{
qcut(x, n, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Number of bins to divide x into (e.g., 3 for tertiles, 5 for quintiles)}

\item{...}{Additional arguments to cut}
}
\description{
Quantile-based cut
}
\examples{
rnorm(9) |> qcut(3) |> table()
}
