% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd.r
\name{qd}
\alias{qd}
\title{Compute a quadratic form (x'Ax)}
\usage{
qd(x, A)
}
\arguments{
\item{x}{A vector}

\item{A}{A matrix compatible with x}
}
\description{
Compute a quadratic form (x'Ax)
}
\examples{
x <- rnorm(10)
A <- matrix(rnorm(100), 10, 10)
qd(x, A)
}
