% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qe.r
\name{qe}
\alias{qe}
\title{Solutions to the quadratic equation}
\usage{
qe(a, b, c)
}
\arguments{
\item{a, b, c}{Coefficients: ax^2 + bx + c = 0}
}
\description{
Some extra effort is dedicated to making the solution stable even when a is close to zero.
}
\examples{
qe(1, 3, -4)  # -4, 1
qe(0, 1, 2)   # Just -2

}
