% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_new.r
\name{rank_new}
\alias{rank_new}
\title{Apply rank transformation to new data}
\usage{
rank_new(x, new, ...)
}
\arguments{
\item{x}{A vector or matrix to which a rank transformation was applied}

\item{new}{New data to be ranked againt the old}

\item{...}{Further arguments to \code{rank()}}
}
\description{
Apply rank transformation to new data
}
\examples{
x <- rnorm(100)
rank_new(x, 1)
rank_new(x, 4)
X <- matrix(rnorm(100*5), 100, 5)
rank_new(X, rbind(1:5, -2:2))
}
