% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasym.r
\name{rasym}
\alias{rasym}
\title{Generate asymptotic sequence of random normal sums}
\usage{
rasym(n, N, len = N, mu = 0, mean = FALSE)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{N}{Maximum sample size.}

\item{len}{Length of sequence; default: all}

\item{mu}{Mean of normal distribution; default: 0}

\item{mean}{Return means (instead of sums?); default: FALSE}
}
\description{
Returns the sum of bernoulli variables with different probabilities.
}
\examples{
rasym(5, 100, 5)
rasym(7, 100, 5, mean=TRUE)

}
