% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbern.R
\name{rbern}
\alias{rbern}
\title{Generate random bernoulli (non-homogeneous binomial) variates}
\usage{
rbern(N, prob)
}
\arguments{
\item{N}{Number of random variates to generate.}

\item{prob}{Vector of probabilities.}
}
\description{
Returns the sum of bernoulli variables with different probabilities.
}
\examples{
rbern(5, c(.3, .6, .1))

}
