% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdex.r
\name{rdex}
\alias{rdex}
\title{Generate random double exponential variates}
\usage{
rdex(n, mean = 0, var = 1)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{mean}{Mean; default: 0}

\item{var}{Variance; default: 1}
}
\description{
Generate random double exponential variates
}
\examples{
rdex(10)
a <- rdex(10000, mean = 3, var = 4)
var(a)

}
