% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdir.r
\name{rdir}
\alias{rdir}
\title{Generate random variates from the Dirichlet distribution}
\usage{
rdir(n, alpha)
}
\arguments{
\item{n}{Number of random variates to generate (single integer)}

\item{alpha}{Concentration parameter (vector of positive numbers)}
}
\value{
An \code{n} by \code{d} matrix, where \code{d} is the length of \code{alpha}.
}
\description{
Generate random variates from the Dirichlet distribution
}
\examples{
rdir(10, 1:4)
}
