% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpower.r
\name{rpower}
\alias{rpower}
\alias{rsamsize}
\alias{rpower_beta}
\title{Power for regression (or any continuous outcome)}
\usage{
rpower(rsq, n, k = 1, p = k, alpha = 0.05)

rsamsize(rsq, k = 1, p = k, power = 0.8, alpha = 0.05)

rpower_beta(rsq, beta)
}
\arguments{
\item{rsq}{Fraction of variance explained by the features to be tested}

\item{n}{Sample size}

\item{k}{Number of parameters to test}

\item{p}{Number of parameters in model (default: k)}

\item{alpha}{Type 1 error rate}

\item{power}{Desired power}
}
\description{
Power for regression (or any continuous outcome)
}
\examples{
rpower(0.05, 40)
rpower(0.05, 40, 5)
rpower(0.05, 40, 5, 20)
rpower(0.05, 40, 5, 20)
rsamsize(0.05)

rpower_beta(0.1)
rpower_beta(beta=0.13)
}
