% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_examples.R
\name{run_ex}
\alias{run_ex}
\title{Runs R code in testing folder}
\usage{
run_ex()
}
\description{
Tests are expected to be in a directory called \code{tests}.  Intended to be used as
an Rstudio addin (which is why it doesn't accept arguments).
}
