% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{Runs R code in testing folder}
\usage{
run_tests()
}
\description{
Tests are expected to be in a directory called \code{tests}.  Intended to be used as
an Rstudio addin (which is why it doesn't accept arguments).
}
