% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_share.r
\name{s3_share}
\alias{s3_share}
\title{Share Amazon S3 file}
\usage{
s3_share(file, out = stdout(), expiration = 168, r)
}
\arguments{
\item{file}{Character vector of files to share.If path is included, it will be included in the download script save location.}

\item{out}{A connection object or character string, as in \verb{[writeLines()]}, to which R commands are to be saved.  If nothing is supplied, output is to console.}

\item{expiration}{How long until URLs expire (in hours)? Numeric.  Default: 168 (1 week).}

\item{r}{Output R code (as opposed to link)? By default, the function will return links if \code{out=stdout()} and R code if \code{out} otherwise. However, this can be overridden by explicitly setting r to be \code{TRUE} or \code{FALSE}.}
}
\description{
Shares Amazon S3 file(s) and returns either a url or R script.  The latter is particularly useful for batch downloads and specifying file save paths.
}
\details{
Requires aws command line functionality.
}
\examples{
\dontrun{
s3_share('~/web/pdf/BrehenyCV.pdf')
s3_share(c('~/web/pdf/BrehenyCV.pdf', '~/web/pdf/mnet.pdf'), r=TRUE)
}
}
