% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-gather.R
\name{save_gather}
\alias{save_gather}
\alias{Bsave}
\alias{Bgather}
\title{Save and combine (gather) results from batched R sessions}
\usage{
Bsave(res, suffix = "")

Bgather(suffix = "", safe = FALSE)
}
\arguments{
\item{res}{Results to save}

\item{suffix}{By default, files are named 'tmp1.rds', 'tmp2.rds', and so on.  If \code{suffix='a'}, then files will be named 'tmp1a.rds', 'tmp2a.rds', etc.  Furthermore, \code{Bgather()} will save results to '2018-11-07a.rds'.}

\item{safe}{By default, files are named 'tmp1.rds', 'tmp2.rds', and so on.  If \code{suffix='a'}, then files will be named 'tmp1a.rds', 'tmp2a.rds', etc.}
}
\description{
Save and combine (gather) results from batched R sessions
}
\examples{
\dontrun{
A <- rnorm(100)
B <- matrix(rnorm(100), ncol=10, nrow=10)
C <- array(rnorm(1000), dim=c(10,10,10))

Bsave(A)
Bsave(A, "A")

Bsave(list(A=A,B=B,C=C), "B")
Bsave(list(A=A,B=B,C=C), "C")}
}
