% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{save_plot}
\alias{save_plot}
\title{Save current base plot as pdf/png}
\usage{
save_plot(file, ...)
}
\arguments{
\item{file}{File path where the plot will be saved (string)}

\item{...}{Additional arguments to graphics device}
}
\description{
See \code{ggplot2::ggsave()} for saving grid plots.
}
\examples{
plot(1:10)
\dontrun{
# Typical journal options:
save_plot("test.png", width = 3.5, height = 3.5, units = "in", res = 600)
}
}
