% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_revlog.r
\name{scale_revlog}
\alias{scale_revlog}
\alias{scale_x_revlog}
\alias{scale_y_revlog}
\title{Provides a reverse log scale to ggplot; especially useful for p-values}
\usage{
scale_x_revlog(base = exp(1), ...)

scale_y_revlog(base = exp(1), ...)
}
\arguments{
\item{base}{Base of log; default is natural log}

\item{...}{Additional arguments to scale_x_continuous or scale_x_continuous}
}
\description{
Provides a reverse log scale to ggplot; especially useful for p-values
}
\examples{
library(ggplot2)
DF <- data.frame(x = 1:3, y = c(0.1, 0.01, 0.001))
ggplot(DF, aes(x, y)) +
  geom_point() +
  scale_y_revlog(10)

}
