% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_url.r
\name{signed_url}
\alias{signed_url}
\title{Share Amazon CloudFront file via presigned url}
\usage{
signed_url(file, out = stdout(), expiration = 365, r)
}
\arguments{
\item{file}{Character vector of files to share. If path is included,
it will be included in the download script save location.}

\item{out}{A connection object or character string, as in
\verb{[writeLines()]}, to which R commands are to be saved.  If nothing is
supplied, output is to console.}

\item{expiration}{How long until URLs expire (in days)? Numeric.
Default: 365.}

\item{r}{Output R code (as opposed to link)? By default, the
function will return links if \code{out=stdout()} and R code if \code{out}
otherwise. However, this can be overridden by explicitly setting r to be
\code{TRUE} or \code{FALSE}.}
}
\value{
Either a url or R script. The latter is particularly useful for
batch downloads and specifying file save paths.
}
\description{
Shares Amazon CloudFront file. Currently only set up for teaching files;
see \code{bin/signed-url}.
}
\examples{
\dontrun{
signed_url("~/pdf/teaching/7110-f23/dk-1.pdf")
signed_url("~/pdf/teaching/7110-f23/dk-*.pdf", r = TRUE)
}
}
