% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent_load.r
\name{silent_load}
\alias{silent_load}
\title{Run \code{devtools::load_all()} silently}
\usage{
silent_load()
}
\description{
Intended to be used as an Rstudio addin (which is why it doesn't accept
arguments).
}
