% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.r
\name{soft}
\alias{soft}
\title{Soft-thresholding operator}
\usage{
soft(x, l)
}
\arguments{
\item{x}{The thing to be thresholded}

\item{l}{The threshold; must either be length 1 or same langth as x}
}
\description{
Soft-thresholding operator
}
\examples{
soft(-10:10, 2.5)
soft(rep(5.5, 10), 1:10)
soft(matrix(1:30, ncol=3), 10)

}
