% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.dir.r
\name{source.dir}
\alias{source.dir}
\title{Source all .R files in a directory}
\usage{
source.dir(directory)
}
\arguments{
\item{directory}{The directory to source}
}
\description{
Source all .R files in a directory
}
\examples{
\dontrun{
source.dir('R')
}

}
