% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2df.r
\name{table2dt}
\alias{table2dt}
\alias{table2df}
\title{Convert a table to a (disaggregated) data frame / table}
\usage{
table2dt(Tab)

table2df(Tab)
}
\arguments{
\item{Tab}{The table}
}
\description{
Note that as.data.frame can do this, but you get aggregated rows with a \code{Freq} column
}
\section{Functions}{
\itemize{
\item \code{table2df}: Convert a table to a (disaggregated) data frame
}}

\examples{
table2dt(Titanic)
head(table2df(Titanic))

}
