% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textPlot.R
\name{textPlot}
\alias{textPlot}
\title{Scatter plot, but with text labels}
\usage{
textPlot(x, y, xlab, ylab, ...)
}
\arguments{
\item{x}{x positions (named); alternatively, a 2-column matrix with column and row names}

\item{y}{y positions, if x is not a matrix}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{...}{Additional arguments to \code{plot()}}
}
\description{
Scatter plot, but with text labels
}
\examples{
a <- runif(10)
b <- runif(10)
names(a) <- names(b) <- LETTERS[1:10]
textPlot(a, b)

X <- cbind(a=a, b=b)
textPlot(X)

}
