% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpower.R
\name{tpower}
\alias{tpower}
\alias{tsamsize}
\title{Power for t tests, including arbitrary linear hypotheses}
\usage{
tpower(n, delta, lam = c(1, -1), b = c(delta, 0), sd = 1,
  alpha = 0.05, w = rep(1, g), n1, n2)

tsamsize(delta, b = c(delta, 0), w = rep(1, g), power = 0.8,
  upper = 5000, ...)
}
\arguments{
\item{n}{Sample size (total)}

\item{delta}{Effect size}

\item{lam}{Contrast}

\item{b}{Coefficient vector}

\item{sd}{Standard deviation of outcome}

\item{alpha}{Type I error rate}

\item{w}{Weights for unequal allocation (normalized to 1)}

\item{n1}{Sample size for group 1}

\item{n2}{Sample size for group 2}

\item{power}{Desired power}
}
\description{
Power for t tests, including arbitrary linear hypotheses
}
\section{Functions}{
\itemize{
\item \code{tpower}: Calculates power

\item \code{tsamsize}: Calculates sample size
}}

\examples{
tpower(100, 0.5)
tsamsize(0.5)

}
