% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{url}
\alias{url}
\title{Open a url: quietly, in the background, and in a new window}
\usage{
url(x)
}
\arguments{
\item{x}{URL to open}
}
\description{
Open a url: quietly, in the background, and in a new window
}
\examples{
url('www.google.com')
}
