% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-fwf.R
\name{write_fwf}
\alias{write_fwf}
\title{Generate fixed width file in R}
\usage{
write_fwf(dt, width, con = stdout(), justify = "l", replace_na = "NA")
}
\arguments{
\item{dt}{The data to be printed}

\item{width}{Either a single number or a vector of widths}

\item{con}{Connection, as in \code{writeLines()}; default is \code{stdout()}.}

\item{justify}{"l", "r" or something like "lrl" for left, right, left.}

\item{replace_na}{What to print in places of missing values; default: "NA"}
}
\description{
https://gist.github.com/haozhu233/28d1309b58431f4929f78243054f1f58
}
\examples{
dt <- data.frame(a = 1:3, b = NA, c = c('a', 'b', 'c'))
write_fwf(dt, "test.txt", width = c(4, 4, 3))

}
