% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tsv.r
\name{write_tsv}
\alias{write_tsv}
\title{Write a tab-separated file, rounding floats}
\usage{
write_tsv(object, precision = 8, sep = "\\t", ...)
}
\arguments{
\item{object}{A \code{data.table} (or coerceable)}

\item{precision}{The number of digits after the decimal point to display (default: 8)}

\item{sep}{Delimiter (default \verb{\\t})}

\item{...}{Additional options to \code{data.table::fwrite()}}
}
\description{
Write a tab-separated file, rounding floats
}
\examples{
dat <- data.table::data.table(a=rnorm(10), b=rexp(10))
write_tsv(dat, 4)
}
