% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsdt.R
\name{xlsdt}
\alias{xlsdt}
\title{Simple wrapper to read_excel + pipe to data.table}
\usage{
xlsdt(...)
}
\arguments{
\item{...}{Arguments to \code{readxl::read_excel()}}
}
\description{
Simple wrapper to read_excel + pipe to data.table
}
