% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Checks whether two vectors are equal; wrapper for \code{all.equal}}
\usage{
check(x, y, check.attributes = FALSE, ...)
}
\arguments{
\item{x, y}{Vectors to test}

\item{check.attributes}{Default: FALSE}

\item{...}{Additional arguments to \code{all.equal()}}
}
\description{
Checks whether two vectors are equal; wrapper for \code{all.equal}
}
\examples{
.test <- 'Example'
check(1,1)
try(check(1,2), silent=TRUE)
cat(geterrmessage())

}
