% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.r
\name{dot-f}
\alias{dot-f}
\alias{.f}
\alias{.p}
\title{Compact and pipe-friendly wrapper to sprintf}
\usage{
.f(x, digits)

.p(x, digits)
}
\arguments{
\item{x}{A floating point number}

\item{digits}{Number of decimals to display}
}
\description{
\code{.p()} is the same as \code{.f()}, but multiplies by 100 first (for percents).
}
\examples{
runif(1) |> .f(2)
}
