% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.r
\name{.f}
\alias{.f}
\title{Compact and pipe-friendly wrapper to sprintf}
\usage{
.f(x, digits)
}
\arguments{
\item{x}{A floating point number}

\item{digits}{Number of decimals to display}
}
\description{
Compact and pipe-friendly wrapper to sprintf
}
\examples{
runif(1) |> .f(2)
}
