% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epower.r
\name{epower}
\alias{epower}
\alias{esamsize}
\title{Power and sample size for time-to-event data based on exponential closed form solution}
\usage{
epower(n, r, alpha = 0.05, w = c(1, 1), n1, n2, lam, cens)

esamsize(r, w = c(1, 1), alpha = 0.05, power = 0.8, lam, cens)
}
\arguments{
\item{n}{Total sample size.}

\item{r}{Effect size (ratio of hazards/medians).}

\item{alpha}{Type I error rate.}

\item{w}{Controls sample size balance between two groups.  Default: Equal size groups.}

\item{n1, n2}{Sample size in each group.}

\item{lam}{Hazard rate in each group (if specified, overrides \code{r}).}

\item{cens}{Censoring rate.}

\item{power}{Desired power.  Default: 0.8.}
}
\description{
Power is calculated for two-sample studies.  Can be vectorized, but \code{w}, \code{lam}, \code{cens}, must always be length 2.
}
\section{Functions}{
\itemize{
\item \code{epower}: Calculate power

\item \code{esamsize}: Calculate sample size
}}

\examples{
epower(30, 2)
epower(30, lam=c(2,1), cens=c(0.2, 0.5))
epower(seq(30, 100, 10), lam=c(2,1), cens=c(0.2, 0.5))
epower(seq(30, 100, 10), w=c(2,1), lam=c(2,1), cens=c(0.2, 0.5))
epower(30, seq(1.5, 2.5, 0.1))
epower(30, 2, alpha=c(0.01, 0.05, 0.1))
esamsize(1.5)
}
