% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.r
\name{flowchart}
\alias{flowchart}
\alias{flow}
\alias{Flow}
\alias{plot.Flow}
\title{Create a flowchart}
\usage{
\method{plot}{Flow}(x, xm = 0.1, ym = 0.1, hpad = 10, vpad = 10, ...)
}
\arguments{
\item{x}{\code{Flow} object}

\item{xm, ym}{Outer x/y margins of the plot (.2 = 20\% margin on each side); default: 0.1}

\item{hpad, vpad}{Horzontal/verticl padding around text in chart nodes, in mm; default: 10}

\item{...}{For S3 method compatibility}
}
\description{
\code{Flow} is an R6 class used for initializing, constructing and modifying the
information that goes into the flowchart.  The plotting is then handled via the
\code{plot()} function.
}
\examples{
fl <- Flow$new("All children\n(n=935)")
fl$add("Returned form\n(n=511)", parent = 1)
fl$add("Baseline: 0\n(n=447)", parent = 2, x = -1)
fl$add("Baseline: 1 or 2\n(n=14)", parent = 2)
fl$add("Baseline: 3\n(n=50)", parent = 2, x = 1)
fl$add("Initiation\n(n=315)", parent = 3, x = -1)
fl$add("Completion\n(n=276)", parent = 6, x = -1)
fl$add("Completion\n(n=12)", parent = 4)
fl
plot(fl)
plot(fl, hpad = 20, vpad = 20)
plot(fl, xm = 0.25)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n}}{Number of nodes}

\item{\code{Parent}}{Vector of parent nodes}

\item{\code{x}}{Vector of horizontal positions}

\item{\code{Label}}{Vector of labels}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Flow-new}{\code{Flow$new()}}
\item \href{#method-Flow-print}{\code{Flow$print()}}
\item \href{#method-Flow-add}{\code{Flow$add()}}
\item \href{#method-Flow-clone}{\code{Flow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flow-new"></a>}}
\if{latex}{\out{\hypertarget{method-Flow-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new flowchart at its parent node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flow$new(lab, x = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lab}}{Label (node text)}

\item{\code{x}}{Horizontal position (0=center)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flow-print"></a>}}
\if{latex}{\out{\hypertarget{method-Flow-print}{}}}
\subsection{Method \code{print()}}{
Print the current status of the flowchart
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flow$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flow-add"></a>}}
\if{latex}{\out{\hypertarget{method-Flow-add}{}}}
\subsection{Method \code{add()}}{
Add a node to the flowchart
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flow$add(lab, parent, x = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lab}}{Label (node text)}

\item{\code{parent}}{Number (i.e., row) of this node's parent}

\item{\code{x}}{Horizontal position (0=center)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Flow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
