% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbox.R
\name{ggbox}
\alias{ggbox}
\title{Boxplot of a data frame / matrix using ggplot}
\usage{
ggbox(X, xlab = "ind", ylab = "values")
}
\arguments{
\item{X}{Matrix of values}

\item{xlab}{Horizontal axis label (default: ind)}

\item{ylab}{Vertical axis label (default: values)}
}
\description{
Boxplot of a data frame / matrix using ggplot
}
\examples{
ggbox(matrix(rnorm(50), 10, 5))

}
