% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{Simple interface to ggplot2 histogram}
\usage{
gghist(x, density = FALSE, ...)
}
\arguments{
\item{x}{A vector of values to make a histogram out of}

\item{density}{Should y-axis be density? (default: FALSE)}

\item{...}{Further arguments to \code{ggplot2::geom_histogram()}}
}
\description{
Simple interface to ggplot2 histogram
}
\examples{
gghist(rnorm(100))
gghist(runif(100), breaks=seq(0, 1, 0.1))
gghist(rnorm(100), density=TRUE)
}
