% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_to_mi.r
\name{km_to_mi}
\alias{km_to_mi}
\alias{mi_to_km}
\title{Convert between min/km pace and min/mi pace}
\usage{
km_to_mi(m, s = 0, half = FALSE)

mi_to_km(m, s = 0, half = FALSE)
}
\arguments{
\item{m}{Minutes}

\item{s}{Seconds (default: 0)}

\item{half}{Per half km? (default: FALSE)}
}
\description{
Calculation is simple, but the minutes/seconds math somewhat cumbersome
}
\examples{
km_to_mi(3, 50)
mi_to_km(6)
}
