% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.r
\name{pca_plot}
\alias{pca_plot}
\title{Make a PCA/tSNE/UMAP plot}
\usage{
pca_plot(
  X,
  grp,
  txt = FALSE,
  method = c("pca", "tsne", "umap"),
  dims = 2,
  gg = TRUE,
  ellipse = FALSE,
  legend,
  repel = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{X}{A numeric matrix}

\item{grp}{Optional grouping factor to color instances by}

\item{txt}{Use rownames as labels instead of dots (default: false)}

\item{method}{How to carry out dimension reduction (pca/tsne/umap)}

\item{dims}{Dimensions to return (for tsne)}

\item{gg}{Use ggplot2? (default: true)}

\item{ellipse}{Draw mvn-ellipses around groups? (default: false, only available for ggplot)}

\item{legend}{Include a legend (default: true)}

\item{repel}{If \code{gg=TRUE} and \code{txt=TRUE}, use \code{ggrepel::geom_text_repel()}? (default: false)}

\item{plot}{Create plot? (default: true)}

\item{...}{Further arguments to \code{plot()}}
}
\value{
If base plot, returns PCA object invisibly. If ggplot, returns the plot object.
}
\description{
The function allows the passing of some parameters to Rtsne/umap:
\itemize{
\item Rtsne: perplexity
\item umap: nn (number of neighbors)
}
}
\examples{
# Plots of the iris data
pca_plot(iris[,1:4])
pca_plot(iris[,1:4], grp=iris[,5])
pca_plot(iris[,1:4], grp=iris[,5], ellipse=TRUE)
pca_plot(iris[,1:4], grp=iris[,5], method='tsne')
pca_plot(iris[,1:4], grp=iris[,5], method='umap')

# What pca_plot returns if plot=FALSE
head(pca_plot(iris[,1:4], method='pca', plot=FALSE))
head(pca_plot(iris[,1:4], method='tsne', dims=3, plot=FALSE))
head(pca_plot(iris[,1:4], method='umap', dims=3, plot=FALSE))

# Plots of the USArrests data
pca_plot(USArrests, txt=TRUE)
pca_plot(USArrests, txt=TRUE, repel=TRUE)
pca_plot(USArrests, txt=TRUE, grp=rep(LETTERS[1:5], each=10))
pca_plot(USArrests, txt=TRUE, method='tsne', perplexity=10)
pca_plot(USArrests, txt=TRUE, method='umap', nn=6)
}
