% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot}
\alias{Plot}
\title{Customized plotting functions}
\usage{
Plot(obj, ...)
}
\arguments{
\item{obj}{Object to be plotted}

\item{...}{Additional arguments to underlying plot function}
}
\description{
Customized plotting functions
}
\examples{
# KM curves
data(veteran, package='survival')
fit <- survival::survfit(survival::Surv(time, status) ~ celltype, veteran)
Plot(fit)

# Trees
data(kyphosis, package='rpart')
fit <- rpart::rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)
Plot(fit)
fit <- party::ctree(Kyphosis ~ Age + Number + Start, data = kyphosis)

}
