% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-hd.R
\name{power_hd}
\alias{power_hd}
\title{Sample size / power calculator for high-dimensional testing}
\usage{
power_hd(n, p, d, m, FDR = 0.1)
}
\arguments{
\item{n}{Sample size (per group)}

\item{p}{Number of features (genes)}

\item{d}{Number of differential features (differentially expressed genes)}

\item{m}{Minimum difference among differential features, in terms of mean difference divided by SD (effect size)}

\item{FDR}{Target FDR}
}
\description{
Sample size / power calculator for high-dimensional testing
}
\examples{
power_hd(10, 1000, 100, 1)
power_hd(seq(10, 100, by=10), 1000, 100, 1)

}
