% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-hd.R
\name{power_hd}
\alias{power_hd}
\title{Power calculator for high-dimensional testing}
\usage{
power_hd(n, p, d, m, FDR = 0.1)
}
\arguments{
\item{n}{Sample size (per group)}

\item{p}{Number of features (genes)}

\item{d}{Number of differential features (differentially expressed genes)}

\item{m}{Minimum difference among differential features, in terms of mean difference divided by SD (effect size)}

\item{FDR}{Target FDR}
}
\value{
A list containing two items:
\itemize{
\item \code{Power}: The power to detect a given feature (function assumes this is the same for all features)
\item \code{Hits}: The total number of expected discoveries (power multiplied by number of differential features)
}
}
\description{
The function sets up a root-finding problem in which the appropriate per-test cutoff alpha is found that satisfies the supplied FDR (in expectation), then reports overall power for tests carried out at that alpha level.  Vectorized over n, but nothing else.
}
\examples{
power_hd(10, 1000, 100, 1)
power_hd(seq(10, 100, by=10), 1000, 100, 1)

}
