% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.r
\name{r2}
\alias{r2}
\title{Nagelkerke R2}
\usage{
r2(fit)
}
\arguments{
\item{fit}{Output of \code{\link[=glm]{glm()}}}
}
\value{
A named vector:
\itemize{
\item \strong{mf:} McFadden \eqn{R^2}
\item \strong{cs:} Cox-Snell \eqn{R^2}
\item \strong{ng:} Nagelkerke \eqn{R^2}
}
}
\description{
Nagelkerke R2
}
\examples{
fit <- glm(am ~ hp + wt, data = mtcars, family = binomial)
r2(fit)
}
