% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samsize_pcc.R
\name{samsize_pcc}
\alias{samsize_pcc}
\title{Sample size for classifier development}
\usage{
samsize_pcc(tolerance, stdfc, p = 0.5, nfeat, dfeat = 20)
}
\arguments{
\item{tolerance}{Sample size is found such that the absolute value of \eqn{PCC(\infty) - PCC(n)} is less than \code{tolerance}.}

\item{stdfc}{Standardized fold change (difference in means on log scale divided by standard deviation)}

\item{p}{Proportion of less common class (default 0.5)}

\item{nfeat}{Number of features}

\item{dfeat}{Number of differential features}
}
\value{
Object of class `power.htest``, a list of the arguments augmented with method and note elements.
}
\description{
Targets the probability of correct classification (PCC) to within a certain tolerance of the optimal PCC.
}
\details{
Loosely based on the function \code{MKmisc::ssize.pcc()}, but more accurate.  I think it is also more accurate than the results presented by Dobbin (2008), but I'm not 100% clear on their method when n1 != n2.  When n1=n2, all methods agree.
}
\examples{
samsize_pcc(0.1, 2, 0.3, 22000)

}
