% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-fwf.r
\name{write_fwf}
\alias{write_fwf}
\title{Generate fixed width file in R}
\usage{
write_fwf(dt, width, con = stdout(), align = "l", na = "NA", col.names = TRUE)
}
\arguments{
\item{dt}{The data to be printed}

\item{width}{Either a single number or a vector of widths}

\item{con}{Connection, as in \code{writeLines()}; default is \code{stdout()}.}

\item{align}{"l", "r" or something like "lrl" for left, right, left.}

\item{na}{What to print in places of missing values; default: "NA"}

\item{col.names}{Print column names?  Default: TRUE}
}
\description{
https://gist.github.com/haozhu233/28d1309b58431f4929f78243054f1f58
}
\examples{
dt <- data.frame(a = 1:3, b = NA, c = c('a', 'b', 'c'))
write_fwf(dt, width = c(4, 4, 3))

}
